requirements.txt
langchain-community
python-dotenv
google-generativeai
